+++ common/src/main/java/net/frozenblock/lib/FrozenLibMain.java
@@ -32,6 +32,7 @@ import net.frozenblock.lib.integration.api.ModIntegrations;
 import net.frozenblock.lib.item.api.component.FrozenLibDataComponents;
 import net.frozenblock.lib.item.impl.cooldown.SerializableItemCooldowns;
 import net.frozenblock.lib.item.impl.loot.predicates.FrozenLibLootConditionTypes;
+import net.frozenblock.lib.levelgen.attribute.impl.FrozenLibEnvironmentAttributes;
 import net.frozenblock.lib.levelgen.biome.impl.modifications.BiomeModificationImpl;
 import net.frozenblock.lib.levelgen.feature.api.FrozenLibFeatures;
 import net.frozenblock.lib.levelgen.feature.impl.blockpredicates.FrozenLibBlockPredicateTypes;
@@ -99,6 +100,7 @@ public final class FrozenLibMain {
 		SoundPredicate.init();
 		MovingSoundTypes.init();
 		FrozenLibParticleTypes.init();
+		FrozenLibEnvironmentAttributes.init();
 		FrozenLibRuleBlockEntityModifiers.init();
 		FrozenLibStructureProcessorTypes.init();
 		FrozenLibStructurePoolElementTypes.init();