+++ common/src/main/java/net/frozenblock/lib/entity/impl/variant/FrozenLibSpawnConditions.java
@@ -2,19 +2,22 @@ package net.frozenblock.lib.entity.impl.variant;
 
 import com.mojang.serialization.MapCodec;
 import net.frozenblock.lib.FrozenLibConstants;
-import net.minecraft.core.Registry;
-import net.minecraft.core.registries.BuiltInRegistries;
+import net.frozenblock.lib.platform.api.registry.FrozenDeferredRegister;
+import net.minecraft.core.registries.Registries;
 import net.minecraft.world.entity.variant.SpawnCondition;
 import org.jetbrains.annotations.ApiStatus;
 
 @ApiStatus.Internal
 public final class FrozenLibSpawnConditions {
+	private static final FrozenDeferredRegister<MapCodec<? extends SpawnCondition>> REGISTER = FrozenDeferredRegister.create(
+		Registries.SPAWN_CONDITION_TYPE,
+		FrozenLibConstants.MOD_ID
+	);
 
-	public static void init() {
-		register("config", ConfigCheck.MAP_CODEC);
+	static {
+		REGISTER.register("config", () -> ConfigCheck.MAP_CODEC);
+		REGISTER.register();
 	}
 
-	private static void register(String name, MapCodec<? extends SpawnCondition> codec) {
-		Registry.register(BuiltInRegistries.SPAWN_CONDITION_TYPE, FrozenLibConstants.id("config"), codec);
-	}
+	public static void init() {}
 }